+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+ - Job restoration - Fix script limiting it to 10 instead of 10'000
+
 ## [2.3.7] - 2019-10-10
 
 ### Changed